Modified ring buffer and Audio Pipeline as per mic output

+++ esp32_firmware/components/amp/amp.h
@@ -1,9 +1,7 @@
 #ifndef AMP_H
 #define AMP_H
-#include <inttypes.h>
-#include <stdint.h>
 
 void amp_init(void);
-void amp_write_block(int32_t *buffer, int size);
+void amp_write_block(float *buffer, int size);
 
 #endif
